requirements.txt
asgiref==3.9.2
Django==5.2.6
django-cors-headers==4.4.0
django-etc==1.4.0
django-hitcount==1.3.5
django-jazzmin==3.0.1
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
